gmaajid-caddi-worksample/pyproject.toml
[project]
name = "caddi-worksample"
version = "0.1.0"
description = "CADDi Solutions Work Sample - Supply Chain Intelligence for Hoth Industries"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pandas>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "anthropic>=0.40.0",
    "cleanco>=2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
caddi-cli = "rag.cli:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rag*", "src*"]
